Document generation and notification services
1) Set up a new module to generate permit and receipt document. The module is imported by application and permit modules to generate receipts and issue, void or revoke permit documents. The new functions in the new module will also allow to reuse the methods for Cron Job ( Doc Gen failure scenarios)
2) Rewoked email notification resend as per latest discussion. Removed overloading of resend.
3) Refactor and code clean up
