+++ vehicles/src/common/enum/notification-type.enum.ts
@@ -1,0 +1,4 @@
+export enum NotificationType {
+  EMAIL_PERMIT = 'EMAIL_PERMIT',
+  EMAIL_RECEIPT = 'EMAIL_RECEIPT',
+}
