+++ vehicles/src/modules/common/dto/response/read-notification.dto.ts
@@ -1,4 +1,7 @@
+import { NotificationType } from '../../../../common/enum/notification-type.enum';
+
 export class ReadNotificationDto {
+  notificationType?: NotificationType;
   message: string;
   transactionId: string;
 }
