+++ vehicles/src/modules/permit-application-payment/permit/permit.module.ts
@@ -8,11 +8,13 @@ import { PermitData } from './entities/permit-data.entity';
 import { Permit } from './entities/permit.entity';
 import { PermitType } from './entities/permit-type.entity';
 import { PaymentModule } from '../payment/payment.module';
+import { PermitReceiptDocumentModule } from '../permit-receipt-document/permit-receipt-document.module';
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([Permit, PermitData, PermitType]),
     PaymentModule,
+    PermitReceiptDocumentModule,
   ],
   controllers: [PermitController],
   providers: [PermitService, PermitProfile],
